requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
Pillow>=7.1.1
gevent>=20.6.2
requests>=2.7.4
torch==1.5.1+cpu
torchvision==0.6.1+cpu

